refactor: 지도 컴포넌트 분리
